backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
spotipy==2.24.0
openai==1.56.2
python-dotenv==1.0.1